Nuclei Template Exploit CVE-2022-1388 BIG-IP iControl REST
CVE-2022-1388 is an authentication bypass vulnerability in the REST 
component of BIG-IP’s iControl API that was assigned a CVSSv3 
score of 9.8. The iControl REST API is used for the management and 
configuration of BIG-IP devices. CVE-2022-1388 could be exploited 
by an unauthenticated attacker with network access to the management 
port or self IP addresses of devices that use BIG-IP. Exploitation would 
allow the attacker to execute arbitrary system commands, create and 
delete files and disable services.